refactor: refactor main.ts and main-lambda-boostrap.ts for deploy Lambda AWS
