+++ campusmap-server/src/main-lambda-bootstrap.ts
@@ -1,59 +1,30 @@
 // main-lambda-bootstrap.ts
-// Bootstrap file for Lambda, exports a function to create the Nest app
 
-import { NestFactory , HttpAdapterHost } from '@nestjs/core';
+import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-import { AppLoggerService } from './common/logger/app-logger.service';
-import { AllExceptionsFilter } from './common/filters/all-exceptions.filter';
+import { setupApp } from './setup-app'; // <-- Importar la función común
+import { INestApplication } from '@nestjs/common';
 
-export async function createApp() {
-   const app = await NestFactory.create(AppModule);
-   const appLogger = await app.resolve(AppLoggerService);
-   app.useLogger(appLogger);
-   const adapterHost = app.get(HttpAdapterHost);
-   app.useGlobalFilters(new AllExceptionsFilter(adapterHost, appLogger));
-   app.setGlobalPrefix('api/v1');
- 
-   if (!process.env.CORS_ALLOW) {
-     appLogger.error(
-       'Error: La variable de entorno CORS_ALLOW no está definida',
-     );
-     process.exit(1);
-   }
- 
-   const corsOrigins = process.env.CORS_ALLOW.split(',')
-     .map((origin) => origin.trim())
-     .filter((origin) => origin.length > 0);
- 
-   if (corsOrigins.length === 0) {
-     appLogger.error(
-       'Error: La variable CORS_ALLOW está vacía o no contiene orígenes válidos',
-     );
-     process.exit(1);
-   }
- 
-   app.enableCors({
-     origin: corsOrigins,
-     methods: ['GET', 'POST', 'PUT', 'PATCH', 'DELETE', 'OPTIONS'],
-     credentials: true,
-     allowedHeaders: ['Content-Type', 'Authorization', 'Accept'],
-   });
- 
-   
- 
-   const config = new DocumentBuilder()
-     .setTitle('CampusMap Server')
-     .setDescription('API de CampusMap')
-     .setVersion('1.0.0')
-     .build();
-   const document = SwaggerModule.createDocument(app, config);
-   SwaggerModule.setup('api/docs', app, document, {
-     swaggerOptions: { persistAuthorization: true },
-   });
-   const port = Number(process.env.PORT ?? 3001);
-   await app.listen(port);
-   appLogger.log(`Servidor escuchando en http://localhost:${port}`);
-   appLogger.log(`Documentación API en http://localhost:${port}/api/docs`);
-   appLogger.log(`CORS habilitado para: ${corsOrigins.join(', ')}`);
+// Nota: Mantener el nombre de exportación como 'createApp' para que el handler de Lambda lo use.
+export async function createApp(): Promise<INestApplication> {
+  console.log('[Lambda] Iniciando creación de la app NestJS...');
+  
+  try {
+    // Nota: El logger: false es útil en Lambda para evitar doble logging al inicio
+    const app = await NestFactory.create(AppModule, { logger: false });
+    console.log('[Lambda] App NestJS creada. Configurando lógica común...');
+    
+    // Lógica de configuración común
+    await setupApp(app); 
+
+    // IMPORTANTE: En Lambda, se usa app.init() en lugar de app.listen()
+    await app.init(); 
+    
+    console.log('[Lambda] App NestJS inicializada correctamente.');
+    return app;
+    
+  } catch (err) {
+    console.error('[Lambda] Error inicializando la app NestJS:', err);
+    throw err;
+  }
 }
